Create C or D.py
Mr.Ram has been given a problem kindly help him to solve it. The input of the program is either 0 or 1. IF 0 is the input he should display "C" if 1 is the input it should display "D".There is a constraint that Mr. Ram should use either logical operators or arithmetic operators to solve the problem, not anything else.
Hint:
Use ASCII values of C and D.
Input Format:
An integer x, 0<=x<=1. .
Output Format:
output a single character "C" or "D"depending on the value of x.
Input 1:
0
Output 1:
C
Input 2:
1
Output 1:
D